Move main package to top-level
Jussi Virtanen pointed out that the installation one-liner looks like a
typo because of the "capstan/capstan" postfix:

  go get github.com/cloudius-systems/capstan/capstan

Fix that by moving main package to top level.
