+++ util/parser.go
@@ -1,4 +1,4 @@
-package capstan
+package util
 
 import (
 	"fmt"
